refractor(model): separe default model from trait definition
